\name{ksmooth}
\alias{ksmooth}
\title{ksmooth with S4-Class value}
\description{
changes the function \code{\link[stats]{ksmooth}} to S4-Class value
}
\usage{
ksmooth(...)
}
\arguments{
  \item{\dots}{ usage described in the original help file. }
}

\value{
  an object of class \code{"ksmooth"}.
}

\seealso{for the class \code{"ksmooth"} see \code{\link[=ksmooth-class]{ksmooth}}. For details of the usage, please see \code{\link[stats]{ksmooth}}.}

\keyword{regression}
\keyword{smooth }
