\name{locfit-class}
\docType{class}
\alias{locfit-class}


\title{Class "locfit"}
\description{ S4-Class for the old class \code{locfit}}
\section{Methods}{
  \describe{
  	\item{monoproc}{\code{signature(fit = "locfit", bandwidth= "numeric", xx= "missing", dir= "missing")}: monotonizing method.}
  	\item{monoproc}{\code{signature(fit = "locfit", bandwidth= "numeric", xx= "missing", dir= "character")}: monotonizing method.}
  	\item{monoproc}{\code{signature(fit = "locfit", bandwidth= "numeric", xx= "numeric", dir= "missing")}: monotonizing method.}
  	\item{monoproc}{\code{signature(fit = "locfit", bandwidth= "numeric", xx= "list", dir= "character")}: monotonizing method.}
	}
}


\seealso{
\code{\link[locfit]{locfit}}
}

\keyword{classes}
