\name{locpoly}
\alias{locpoly}
\title{locpoly with S4-Class value}
\description{
  changes the function \code{\link[KernSmooth]{locpoly}} to S4-Class value
}
\usage{
locpoly(...)
}
\arguments{
  \item{\dots}{  usage described in the original help file. }
}

\value{
an object of class \code{"locpoly"}
}




\seealso{ for the class \code{"locpoly"} see \code{\link[=locpoly-class]{locpoly}}. 
For the usage, see  \code{\link[KernSmooth]{locpoly}}.}


\keyword{ regression}
\keyword{ smooth }
