\name{loess-class}
\docType{class}
\alias{loess-class}


\title{Class "loess"}
\description{ S4-Class for the old class \code{loess}}
\section{Methods}{
  \describe{
  	\item{monoproc}{\code{signature(fit = "loess", bandwidth= "numeric", xx= "missing", dir= "missing")}: monotonizing method.}
  	\item{monoproc}{\code{signature(fit = "loess", bandwidth= "numeric", xx= "numeric", dir= "missing")}: monotonizing method.}
  	}
}


\seealso{
\code{\link[stat]{loess}}
}

\keyword{classes}
